♻️ Complete server module decomposition
Finish the remaining phases of the server functional refactor by extracting session core state, Claude hook event handlers, mission config helpers, persistence write families, and Codex config support modules.\n\nThis keeps the mutable shell / functional core split clearer, reduces god files across runtime and connector boundaries, and updates the tracked plan to reflect the completed architecture work.
