+++ orbitdock-server/crates/server/src/connectors/claude_hooks/handler.rs
@@ -11,45 +11,45 @@
 //! ghost sessions from `claude -c` bootstrap processes.
 
 use std::sync::Arc;
-use std::time::Instant;
-
-use serde_json::Value;
 use tracing::warn;
 
 #[path = "routing.rs"]
 mod routing;
+#[path = "session_end.rs"]
+mod session_end;
+#[path = "session_start.rs"]
+mod session_start;
+#[path = "status_events.rs"]
+mod status_events;
+#[path = "subagent_events.rs"]
+mod subagent_events;
 #[path = "subagent_updates.rs"]
 mod subagent_updates;
+#[path = "tool_events.rs"]
+mod tool_events;
+#[path = "transcript_sync.rs"]
+mod transcript_sync;
 
-use orbitdock_protocol::domain_events::AgentType;
-use orbitdock_protocol::{
-  ClientMessage, Provider, SessionControlMode, SessionLifecycleState, SessionStatus, SubagentInfo,
-  SubagentStatus,
-};
+use orbitdock_protocol::ClientMessage;
 
-use crate::domain::sessions::transition::{approval_question, Input};
-use crate::infrastructure::persistence::PersistCommand;
-use crate::runtime::session_commands::SessionCommand;
-use crate::runtime::session_registry::{PendingClaudeSession, PendingHookSession, SessionRegistry};
-use crate::support::session_paths::{claude_transcript_path_from_cwd, project_name_from_cwd};
-use crate::support::session_time::chrono_now;
+use crate::runtime::session_registry::SessionRegistry;
 
-use self::routing::{
-  cleanup_claude_shadow_session, resolve_claude_hook_routing, ClaudeHookHandlingOptions,
-  ClaudeHookRoutingDecision,
-};
-use self::subagent_updates::{
-  apply_claude_subagent_update, maybe_sync_transcript_messages, publish_claude_subagent_update,
-  ClaudeSubagentUpdate,
-};
+use self::routing::ClaudeHookHandlingOptions;
+use self::session_end::handle_claude_session_end;
+use self::session_start::handle_claude_session_start;
+use self::status_events::handle_claude_status_event;
+use self::subagent_events::handle_claude_subagent_event;
+#[cfg(test)]
+use self::subagent_updates::{apply_claude_subagent_update, ClaudeSubagentUpdate};
+use self::tool_events::handle_claude_tool_event;
+pub(super) use super::approval;
+#[cfg(test)]
 use super::approval::{
-  classify_permission_request, claude_permission_request_id, extract_plan_from_tool_input,
-  extract_question_from_tool_input, permission_request_matches_snapshot,
-  resolve_pending_approvals_after_tool_outcome, PermissionRequestSnapshotMatch,
+  classify_permission_request, extract_plan_from_tool_input, extract_question_from_tool_input,
 };
-use super::session_materialization::{
-  emit_capabilities_from_transcript, is_codex_rollout_payload, materialize_claude_session,
-};
+pub(super) use super::session_materialization;
+#[cfg(test)]
+use super::session_materialization::is_codex_rollout_payload;
 
 /// Process a Claude hook message.
 /// These handlers never need `client_tx` or `conn_id` — only `state`.
@@ -75,209 +75,24 @@ pub async fn handle_hook_message_with_options(
       terminal_session_id,
       terminal_app,
     } => {
-      // Defensive guard: codex rollout payloads should stay on Codex path.
-      if context_label.as_deref() == Some("codex_cli_rs") {
-        return;
-      }
-
-      // Enhanced Codex filter: reject payloads from Codex CLI sessions
-      if is_codex_rollout_payload(transcript_path.as_deref(), model.as_deref()) {
-        return;
-      }
-
-      match resolve_claude_hook_routing(state, &session_id).await {
-        ClaudeHookRoutingDecision::ManagedDirect {
-          owner_session_id: _,
-        } => {
-          cleanup_claude_shadow_session(state, &session_id, "managed_direct_session").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreShadowedByDirect => {
-          cleanup_claude_shadow_session(state, &session_id, "direct_owner_exists").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreOwnershipLookupFailed => {
-          return;
-        }
-        ClaudeHookRoutingDecision::Passive => {}
-      }
-
-      // If there's a direct Claude session awaiting SDK ID registration, claim it eagerly.
-      // Write goes through PersistCommand only — single mutation path with immutability guard.
-      if let Some(owning_id) = state.find_unregistered_direct_claude_session(&cwd) {
-        state.register_claude_runtime_owner(&session_id, &owning_id);
-        let persisted = state
-          .persist()
-          .send(PersistCommand::SetClaudeSdkSessionId {
-            session_id: owning_id.clone(),
-            claude_sdk_session_id: session_id.clone(),
-          })
-          .await
-          .is_ok();
-        if persisted {
-          return;
-        }
-        state.unregister_claude_runtime_owner(&session_id);
-      }
-
-      // If session already exists (e.g. restored from DB), update it directly
-      // instead of caching as pending.
-      if let Some(existing) = state.get_session(&session_id) {
-        if existing.snapshot().provider == Provider::Codex {
-          return;
-        }
-        if let Some(ref m) = model {
-          existing
-            .send(SessionCommand::ProcessEvent {
-              event: Input::ModelUpdated(m.clone()),
-            })
-            .await;
-        }
-        if transcript_path.is_some() {
-          existing
-            .send(SessionCommand::ProcessEvent {
-              event: Input::TranscriptPathUpdated(transcript_path.clone()),
-            })
-            .await;
-        }
-        let git_info = crate::domain::git::repo::resolve_git_info(&cwd).await;
-        let git_branch = git_info.as_ref().map(|g| g.branch.clone());
-        let git_sha = git_info.as_ref().map(|g| g.sha.clone());
-        let repository_root = git_info.as_ref().map(|g| g.common_dir_root.clone());
-        let is_worktree = git_info.as_ref().is_some_and(|g| g.is_worktree);
-
-        // Use repository root for grouping so worktree sessions group correctly
-        let effective_project_path = repository_root.clone().unwrap_or_else(|| cwd.clone());
-
-        crate::runtime::session_state_transitions::transition_work_status(
-          &existing,
-          &session_id,
-          orbitdock_protocol::WorkStatus::Waiting,
-          Some(orbitdock_protocol::StateChanges {
-            current_cwd: Some(Some(cwd.clone())),
-            git_branch: git_branch.as_ref().map(|b| Some(b.clone())),
-            git_sha: git_sha.as_ref().map(|s| Some(s.clone())),
-            repository_root: repository_root.as_ref().map(|r| Some(r.clone())),
-            is_worktree: if is_worktree { Some(true) } else { None },
-            ..Default::default()
-          }),
-        )
-        .await;
-        existing
-          .send(SessionCommand::ProcessEvent {
-            event: Input::EnvironmentChanged {
-              cwd: Some(cwd.clone()),
-              git_branch: git_branch.clone(),
-              git_sha: git_sha.clone(),
-              repository_root: repository_root.clone(),
-              is_worktree: Some(is_worktree),
-            },
-          })
-          .await;
-        let _ = state
-          .persist()
-          .send(PersistCommand::ClaudeSessionUpsert {
-            id: session_id,
-            project_path: effective_project_path.clone(),
-            project_name: project_name_from_cwd(&effective_project_path),
-            branch: git_branch,
-            model,
-            context_label,
-            transcript_path,
-            source,
-            agent_type,
-            permission_mode,
-            terminal_session_id,
-            terminal_app,
-            forked_from_session_id: None,
-            repository_root,
-            is_worktree,
-            git_sha,
-          })
-          .await;
-        return;
-      }
-
-      // Defer session creation — cache metadata until an actionable hook arrives.
-      state.cache_pending_hook_session(
+      handle_claude_session_start(
+        state,
         session_id,
-        PendingHookSession::Claude(PendingClaudeSession {
-          cwd,
-          model,
-          source,
-          context_label,
-          transcript_path,
-          permission_mode,
-          agent_type,
-          terminal_session_id,
-          terminal_app,
-          cached_at: Instant::now(),
-        }),
-      );
+        cwd,
+        model,
+        source,
+        context_label,
+        transcript_path,
+        permission_mode,
+        agent_type,
+        terminal_session_id,
+        terminal_app,
+      )
+      .await;
     }
 
     ClientMessage::ClaudeSessionEnd { session_id, reason } => {
-      match resolve_claude_hook_routing(state, &session_id).await {
-        ClaudeHookRoutingDecision::ManagedDirect {
-          owner_session_id: _,
-        } => {
-          cleanup_claude_shadow_session(state, &session_id, "managed_direct_session").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreShadowedByDirect => {
-          cleanup_claude_shadow_session(state, &session_id, "direct_owner_exists").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreOwnershipLookupFailed => {
-          return;
-        }
-        ClaudeHookRoutingDecision::Passive => {}
-      }
-
-      // If session was never materialized (ghost from `claude -c`), discard silently.
-      if state.discard_pending_hook_session(Provider::Claude, &session_id) {
-        return;
-      }
-
-      let persist_tx = state.persist().clone();
-
-      if let Some(existing) = state.get_session(&session_id) {
-        if existing.snapshot().provider == Provider::Codex {
-          return;
-        }
-
-        // Extract AI-generated summary from transcript before ending
-        if let Some(transcript_path) = &existing.snapshot().transcript_path {
-          if let Some(summary) =
-            crate::infrastructure::persistence::extract_summary_from_transcript_path(
-              transcript_path,
-            )
-            .await
-          {
-            let _ = persist_tx
-              .send(PersistCommand::SetSummary {
-                session_id: session_id.clone(),
-                summary,
-              })
-              .await;
-          }
-        }
-      }
-
-      let _ = persist_tx
-        .send(PersistCommand::ClaudeSessionEnd {
-          id: session_id.clone(),
-          reason: reason.clone(),
-        })
-        .await;
-
-      if state.remove_session(&session_id).is_some() {
-        let _ = state
-          .list_tx()
-          .send(orbitdock_protocol::ServerMessage::DashboardItemRemoved {
-            session_id: session_id.clone(),
-          });
-      }
+      handle_claude_session_end(state, session_id, reason).await;
     }
 
     ClientMessage::ClaudeStatusEvent {
@@ -303,422 +118,19 @@ pub async fn handle_hook_message_with_options(
       config_source: _,
       config_file_path: _,
     } => {
-      match resolve_claude_hook_routing(state, &session_id).await {
-        ClaudeHookRoutingDecision::ManagedDirect { owner_session_id } => {
-          cleanup_claude_shadow_session(state, &session_id, "managed_direct_session").await;
-
-          if let Some(actor) = state.get_session(&owner_session_id) {
-            let persist_tx = state.persist().clone();
-
-            // Route work_status transitions for managed direct sessions
-            let next_work_status = match hook_event_name.as_str() {
-              "UserPromptSubmit" => Some(orbitdock_protocol::WorkStatus::Working),
-              "Stop" => {
-                let is_question =
-                  actor.last_tool().await.ok().flatten().as_deref() == Some("AskUserQuestion");
-                if is_question {
-                  Some(orbitdock_protocol::WorkStatus::Question)
-                } else {
-                  Some(orbitdock_protocol::WorkStatus::Reply)
-                }
-              }
-              "Notification" => match notification_type.as_deref() {
-                Some("idle_prompt") => {
-                  let is_question =
-                    actor.last_tool().await.ok().flatten().as_deref() == Some("AskUserQuestion");
-                  if is_question {
-                    Some(orbitdock_protocol::WorkStatus::Question)
-                  } else {
-                    Some(orbitdock_protocol::WorkStatus::Reply)
-                  }
-                }
-                _ => None,
-              },
-              "TeammateIdle" => Some(orbitdock_protocol::WorkStatus::Reply),
-              _ => None,
-            };
-
-            if let Some(ws) = next_work_status {
-              use crate::runtime::session_state_transitions::{
-                attention_reason_for_status, transition_work_status,
-              };
-              actor
-                .send(SessionCommand::ProcessEvent {
-                  event: Input::AttentionUpdated {
-                    attention_reason: attention_reason_for_status(ws),
-                    last_tool: None,
-                    pending_tool_name: None,
-                    pending_tool_input: None,
-                    pending_question: None,
-                  },
-                })
-                .await;
-              transition_work_status(&actor, &owner_session_id, ws, None).await;
-
-              crate::runtime::session_registry::flush_and_publish_conversation(
-                &persist_tx,
-                state,
-                &owner_session_id,
-              )
-              .await;
-            }
-
-            // Route summary extraction on Stop
-            if hook_event_name == "Stop" {
-              let snap = actor.snapshot();
-              if snap.summary.is_none() {
-                let derived = cwd
-                  .as_deref()
-                  .and_then(|p| claude_transcript_path_from_cwd(p, &session_id));
-                let tp = snap
-                  .transcript_path
-                  .clone()
-                  .or_else(|| transcript_path.clone())
-                  .or(derived);
-                if let Some(path) = tp {
-                  if let Some(summary) =
-                    crate::infrastructure::persistence::extract_summary_from_transcript_path(&path)
-                      .await
-                  {
-                    actor
-                      .send(SessionCommand::ProcessEvent {
-                        event: Input::SummaryUpdated(summary),
-                      })
-                      .await;
-                  }
-                }
-              }
-            }
-
-            // Route compact_count increment on PreCompact
-            if hook_event_name == "PreCompact" {
-              let _ = persist_tx
-                .send(PersistCommand::ClaudeSessionUpdate {
-                  id: owner_session_id.clone(),
-                  work_status: None,
-                  attention_reason: None,
-                  last_tool: None,
-                  last_tool_at: None,
-                  pending_tool_name: None,
-                  pending_tool_input: None,
-                  pending_question: None,
-                  source: None,
-                  agent_type: None,
-                  permission_mode: None,
-                  active_subagent_id: None,
-                  active_subagent_type: None,
-                  first_prompt: None,
-                  compact_count_increment: true,
-                })
-                .await;
-            }
-
-            // Route last_tool tracking
-            if let Some(ref tool_name) = tool_name {
-              actor
-                .send(SessionCommand::ProcessEvent {
-                  event: Input::AttentionUpdated {
-                    attention_reason: None,
-                    last_tool: Some(tool_name.clone()),
-                    pending_tool_name: None,
-                    pending_tool_input: None,
-                    pending_question: None,
-                  },
-                })
-                .await;
-            }
-          }
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreShadowedByDirect => {
-          cleanup_claude_shadow_session(state, &session_id, "direct_owner_exists").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreOwnershipLookupFailed => {
-          return;
-        }
-        ClaudeHookRoutingDecision::Passive => {}
-      }
-
-      let persist_tx = state.persist().clone();
-      let derived_transcript_path = cwd
-        .as_deref()
-        .and_then(|path| claude_transcript_path_from_cwd(path, &session_id));
-
-      // Resolve full git info from cwd if available
-      let git_info = match cwd.as_deref() {
-        Some(path) => crate::domain::git::repo::resolve_git_info(path).await,
-        None => None,
-      };
-      let git_branch = git_info.as_ref().map(|g| g.branch.clone());
-      let git_sha = git_info.as_ref().map(|g| g.sha.clone());
-      let repository_root = git_info.as_ref().map(|g| g.common_dir_root.clone());
-      let is_worktree = git_info.as_ref().is_some_and(|g| g.is_worktree);
-
-      let actor = if let Some(existing) = state.get_session(&session_id) {
-        if existing.snapshot().provider == Provider::Codex {
-          return;
-        }
-        if cwd.is_some() || git_branch.is_some() || repository_root.is_some() {
-          existing
-            .send(SessionCommand::ProcessEvent {
-              event: Input::EnvironmentChanged {
-                cwd: cwd.clone(),
-                git_branch: git_branch.clone(),
-                git_sha: git_sha.clone(),
-                repository_root: repository_root.clone(),
-                is_worktree: if is_worktree { Some(true) } else { None },
-              },
-            })
-            .await;
-        }
-        existing
-      } else {
-        // Materialize from pending cache or create a fallback session
-        let fallback_cwd = cwd.clone().unwrap_or_else(|| "/unknown".to_string());
-        let actor = materialize_claude_session(
-          &session_id,
-          &fallback_cwd,
-          transcript_path
-            .clone()
-            .or_else(|| derived_transcript_path.clone()),
-          git_info.as_ref(),
-          state,
-          &persist_tx,
-        )
-        .await;
-        actor
-      };
-
-      if transcript_path.is_some() || derived_transcript_path.is_some() {
-        actor
-          .send(SessionCommand::ProcessEvent {
-            event: Input::TranscriptPathUpdated(
-              transcript_path
-                .clone()
-                .or_else(|| derived_transcript_path.clone()),
-            ),
-          })
-          .await;
-      }
-
-      // Use repository root for grouping so worktree sessions group correctly
-      if let Some(cwd) = cwd.clone() {
-        let effective_project_path = repository_root.clone().unwrap_or_else(|| cwd.clone());
-        actor
-          .send(SessionCommand::ProcessEvent {
-            event: Input::EnvironmentChanged {
-              cwd: Some(cwd.clone()),
-              git_branch: git_branch.clone(),
-              git_sha: git_sha.clone(),
-              repository_root: repository_root.clone(),
-              is_worktree: Some(is_worktree),
-            },
-          })
-          .await;
-        let _ = persist_tx
-          .send(PersistCommand::ClaudeSessionUpsert {
-            id: session_id.clone(),
-            project_path: effective_project_path.clone(),
-            project_name: project_name_from_cwd(&effective_project_path),
-            branch: git_branch.clone(),
-            model: None,
-            context_label: None,
-            transcript_path: transcript_path
-              .clone()
-              .or_else(|| derived_transcript_path.clone()),
-            source: None,
-            agent_type: None,
-            permission_mode: permission_mode.clone(),
-            terminal_session_id: None,
-            terminal_app: None,
-            forked_from_session_id: None,
-            repository_root,
-            is_worktree,
-            git_sha,
-          })
-          .await;
-      }
-
-      let next_work_status = match hook_event_name.as_str() {
-        "UserPromptSubmit" => Some(orbitdock_protocol::WorkStatus::Working),
-        "Stop" => {
-          let is_question =
-            actor.last_tool().await.ok().flatten().as_deref() == Some("AskUserQuestion");
-          if is_question {
-            Some(orbitdock_protocol::WorkStatus::Question)
-          } else {
-            Some(orbitdock_protocol::WorkStatus::Reply)
-          }
-        }
-        "Notification" => match notification_type.as_deref() {
-          // Notification events are informational; actionable permission/question
-          // state is driven by PermissionRequest tool hooks.
-          Some("permission_prompt") => None,
-          Some("elicitation_dialog") => None,
-          Some("idle_prompt") => {
-            let is_question =
-              actor.last_tool().await.ok().flatten().as_deref() == Some("AskUserQuestion");
-            if is_question {
-              Some(orbitdock_protocol::WorkStatus::Question)
-            } else {
-              Some(orbitdock_protocol::WorkStatus::Reply)
-            }
-          }
-          _ => None,
-        },
-        "TeammateIdle" => Some(orbitdock_protocol::WorkStatus::Reply),
-        _ => None,
-      };
-
-      if hook_event_name == "UserPromptSubmit" {
-        let _ = persist_tx
-          .send(PersistCommand::ClaudePromptIncrement {
-            id: session_id.clone(),
-            first_prompt: prompt.clone(),
-          })
-          .await;
-
-        // Broadcast first_prompt delta and trigger AI naming
-        if let Some(ref prompt_text) = prompt {
-          actor
-            .send(SessionCommand::ProcessEvent {
-              event: crate::domain::sessions::transition::Input::FirstPromptCaptured(
-                prompt_text.clone(),
-              ),
-            })
-            .await;
-
-          if state.naming_guard().try_claim(&session_id) {
-            crate::support::ai_naming::spawn_naming_task(
-              session_id.clone(),
-              prompt_text.clone(),
-              actor.clone(),
-            );
-          }
-        }
-
-        // Branch freshness: re-resolve git info on every user prompt to catch
-        // branch switches between turns. Cost: ~10ms per user prompt.
-        if let Some(ref prompt_cwd) = cwd {
-          let fresh_info = crate::domain::git::repo::resolve_git_info(prompt_cwd).await;
-          if let Some(ref info) = fresh_info {
-            actor
-              .send(SessionCommand::ProcessEvent {
-                event: Input::EnvironmentChanged {
-                  cwd: Some(prompt_cwd.clone()),
-                  git_branch: Some(info.branch.clone()),
-                  git_sha: Some(info.sha.clone()),
-                  repository_root: Some(info.common_dir_root.clone()),
-                  is_worktree: Some(info.is_worktree),
-                },
-              })
-              .await;
-          }
-        }
-      }
-
-      // On the first prompt, extract capabilities (skills, tools, slash_commands)
-      // from the transcript and broadcast to subscribers. By UserPromptSubmit time
-      // the init system message is always present in the transcript.
-      if hook_event_name == "UserPromptSubmit" {
-        emit_capabilities_from_transcript(&session_id, &actor).await;
-      }
-
-      // On Stop events, try to extract AI-generated summary from transcript.
-      if hook_event_name == "Stop" {
-        let snap = actor.snapshot();
-        if snap.summary.is_none() {
-          let tp = snap
-            .transcript_path
-            .clone()
-            .or_else(|| transcript_path.clone())
-            .or_else(|| derived_transcript_path.clone());
-          if let Some(path) = tp {
-            if let Some(extracted_summary) =
-              crate::infrastructure::persistence::extract_summary_from_transcript_path(&path).await
-            {
-              actor
-                .send(SessionCommand::ProcessEvent {
-                  event: Input::SummaryUpdated(extracted_summary),
-                })
-                .await;
-            }
-          }
-        }
-      }
-
-      if hook_event_name == "PreCompact" {
-        let _ = persist_tx
-          .send(PersistCommand::ClaudeSessionUpdate {
-            id: session_id.clone(),
-            work_status: None,
-            attention_reason: None,
-            last_tool: None,
-            last_tool_at: None,
-            pending_tool_name: None,
-            pending_tool_input: None,
-            pending_question: None,
-            source: None,
-            agent_type: None,
-            permission_mode: None,
-            active_subagent_id: None,
-            active_subagent_type: None,
-            first_prompt: None,
-            compact_count_increment: true,
-          })
-          .await;
-      }
-
-      if let Some(tool_name) = tool_name {
-        actor
-          .send(SessionCommand::ProcessEvent {
-            event: Input::AttentionUpdated {
-              attention_reason: None,
-              last_tool: Some(tool_name),
-              pending_tool_name: None,
-              pending_tool_input: None,
-              pending_question: None,
-            },
-          })
-          .await;
-      }
-
-      if let Some(work_status) = next_work_status {
-        use crate::runtime::session_state_transitions::{
-          attention_reason_for_status, transition_work_status,
-        };
-        actor
-          .send(SessionCommand::ProcessEvent {
-            event: Input::AttentionUpdated {
-              attention_reason: attention_reason_for_status(work_status),
-              last_tool: None,
-              pending_tool_name: None,
-              pending_tool_input: None,
-              pending_question: None,
-            },
-          })
-          .await;
-        transition_work_status(&actor, &session_id, work_status, None).await;
-
-        // Persist permission_mode separately if reported
-        if let Some(ref pm) = permission_mode {
-          actor
-            .send(SessionCommand::ProcessEvent {
-              event: Input::PermissionModeChanged { mode: pm.clone() },
-            })
-            .await;
-        }
-
-        crate::runtime::session_registry::flush_and_publish_conversation(
-          &persist_tx,
-          state,
-          &session_id,
-        )
-        .await;
-      }
-
-      maybe_sync_transcript_messages(&actor, &persist_tx, &options).await;
+      handle_claude_status_event(
+        state,
+        &options,
+        session_id,
+        cwd,
+        transcript_path,
+        hook_event_name,
+        notification_type,
+        tool_name,
+        prompt,
+        permission_mode,
+      )
+      .await;
     }
 
     ClientMessage::ClaudeToolEvent {
@@ -734,391 +146,20 @@ pub async fn handle_hook_message_with_options(
       is_interrupt,
       permission_mode,
     } => {
-      match resolve_claude_hook_routing(state, &session_id).await {
-        ClaudeHookRoutingDecision::ManagedDirect { owner_session_id } => {
-          cleanup_claude_shadow_session(state, &session_id, "managed_direct_session").await;
-          let persist_tx = state.persist().clone();
-
-          match hook_event_name.as_str() {
-            "PreToolUse" => {
-              if let Some(actor) = state.get_session(&owner_session_id) {
-                actor
-                  .send(SessionCommand::ProcessEvent {
-                    event: Input::AttentionUpdated {
-                      attention_reason: None,
-                      last_tool: Some(tool_name.clone()),
-                      pending_tool_name: None,
-                      pending_tool_input: None,
-                      pending_question: None,
-                    },
-                  })
-                  .await;
-              }
-            }
-            "PermissionRequest" => {
-              // For managed direct sessions, the connector's
-              // can_use_tool control_request is the single source
-              // of truth for approvals. The hook PermissionRequest
-              // is redundant — skip approval creation and only
-              // persist supplementary metadata.
-              if let Some(actor) = state.get_session(&owner_session_id) {
-                actor
-                  .send(SessionCommand::ProcessEvent {
-                    event: Input::AttentionUpdated {
-                      attention_reason: None,
-                      last_tool: Some(tool_name),
-                      pending_tool_name: None,
-                      pending_tool_input: None,
-                      pending_question: None,
-                    },
-                  })
-                  .await;
-              }
-            }
-            "PostToolUse" | "PostToolUseFailure" => {
-              // For managed direct sessions, the connector owns
-              // the approval lifecycle via control_response.
-              // Do NOT call resolve_pending_approvals here — it
-              // clears the server queue without sending the CLI
-              // the control_response it's waiting for, causing a
-              // deadlock when parallel tool calls are in flight.
-              // Only persist supplementary metadata (tool count).
-              let _ = persist_tx
-                .send(PersistCommand::ClaudeToolIncrement {
-                  id: owner_session_id.clone(),
-                })
-                .await;
-            }
-            _ => {}
-          }
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreShadowedByDirect => {
-          cleanup_claude_shadow_session(state, &session_id, "direct_owner_exists").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreOwnershipLookupFailed => {
-          return;
-        }
-        ClaudeHookRoutingDecision::Passive => {}
-      }
-
-      let persist_tx = state.persist().clone();
-      let derived_transcript_path = claude_transcript_path_from_cwd(&cwd, &session_id);
-
-      // Resolve full git info from cwd
-      let git_info = crate::domain::git::repo::resolve_git_info(&cwd).await;
-      let git_branch = git_info.as_ref().map(|g| g.branch.clone());
-      let git_sha = git_info.as_ref().map(|g| g.sha.clone());
-      let repository_root = git_info.as_ref().map(|g| g.common_dir_root.clone());
-      let is_worktree = git_info.as_ref().is_some_and(|g| g.is_worktree);
-
-      let actor = if let Some(existing) = state.get_session(&session_id) {
-        if existing.snapshot().provider == Provider::Codex {
-          return;
-        }
-        // Update branch/worktree info if missing
-        if git_branch.is_some() && existing.snapshot().git_branch.is_none() {
-          existing
-            .send(SessionCommand::ProcessEvent {
-              event: Input::EnvironmentChanged {
-                cwd: None,
-                git_branch: git_branch.clone(),
-                git_sha: git_sha.clone(),
-                repository_root: repository_root.clone(),
-                is_worktree: if is_worktree { Some(true) } else { None },
-              },
-            })
-            .await;
-        }
-        existing
-      } else {
-        // Materialize from pending cache or create a fallback session
-        materialize_claude_session(
-          &session_id,
-          &cwd,
-          derived_transcript_path.clone(),
-          git_info.as_ref(),
-          state,
-          &persist_tx,
-        )
-        .await
-      };
-
-      let effective_project_path = repository_root.clone().unwrap_or_else(|| cwd.clone());
-      let _ = persist_tx
-        .send(PersistCommand::ClaudeSessionUpsert {
-          id: session_id.clone(),
-          project_path: effective_project_path.clone(),
-          project_name: project_name_from_cwd(&effective_project_path),
-          branch: git_branch,
-          model: None,
-          context_label: None,
-          transcript_path: derived_transcript_path,
-          source: None,
-          agent_type: None,
-          permission_mode: permission_mode.clone(),
-          terminal_session_id: None,
-          terminal_app: None,
-          forked_from_session_id: None,
-          repository_root,
-          is_worktree,
-          git_sha,
-        })
-        .await;
-
-      match hook_event_name.as_str() {
-        "PreToolUse" => {
-          let snapshot = actor.snapshot();
-          let was_permission = snapshot.work_status == orbitdock_protocol::WorkStatus::Permission;
-          let had_pending_approval = snapshot.pending_approval_id.is_some();
-
-          let question = tool_input
-            .as_ref()
-            .and_then(|value| value.get("question"))
-            .and_then(Value::as_str)
-            .map(|s| s.to_string());
-          let serialized_input = tool_input
-            .as_ref()
-            .and_then(|value| serde_json::to_string(value).ok());
-
-          actor
-            .send(SessionCommand::ProcessEvent {
-              event: Input::AttentionUpdated {
-                attention_reason: None,
-                last_tool: Some(tool_name.clone()),
-                pending_tool_name: if was_permission || had_pending_approval {
-                  None
-                } else {
-                  Some(Some(tool_name.clone()))
-                },
-                pending_tool_input: if was_permission || had_pending_approval {
-                  None
-                } else {
-                  Some(serialized_input)
-                },
-                pending_question: if was_permission || had_pending_approval {
-                  None
-                } else {
-                  Some(question)
-                },
-              },
-            })
-            .await;
-          actor
-            .send(SessionCommand::ApplyDelta {
-              changes: Box::new(orbitdock_protocol::StateChanges {
-                work_status: Some(orbitdock_protocol::WorkStatus::Working),
-                last_activity_at: Some(chrono_now()),
-                ..Default::default()
-              }),
-              persist_op: None,
-            })
-            .await;
-
-          if let Some(ref pm) = permission_mode {
-            actor
-              .send(SessionCommand::ProcessEvent {
-                event: Input::PermissionModeChanged { mode: pm.clone() },
-              })
-              .await;
-          }
-        }
-        "PostToolUse" => {
-          resolve_pending_approvals_after_tool_outcome(
-            &actor,
-            &persist_tx,
-            &session_id,
-            "approved",
-            orbitdock_protocol::WorkStatus::Working,
-          )
-          .await;
-
-          let _ = persist_tx
-            .send(PersistCommand::ClaudeToolIncrement {
-              id: session_id.clone(),
-            })
-            .await;
-          // Clear pending attention fields after tool completion
-          actor
-            .send(SessionCommand::ProcessEvent {
-              event: Input::AttentionUpdated {
-                attention_reason: None,
-                last_tool: None,
-                pending_tool_name: Some(None),
-                pending_tool_input: Some(None),
-                pending_question: Some(None),
-              },
-            })
-            .await;
-
-          // Broadcast permission_mode changes (e.g. EnterPlanMode sets "plan",
-          // ExitPlanMode restores "default") so clients update immediately.
-          if let Some(ref pm) = permission_mode {
-            actor
-              .send(SessionCommand::ProcessEvent {
-                event: Input::PermissionModeChanged { mode: pm.clone() },
-              })
-              .await;
-          }
-          actor
-            .send(SessionCommand::ApplyDelta {
-              changes: Box::new(orbitdock_protocol::StateChanges {
-                work_status: Some(orbitdock_protocol::WorkStatus::Working),
-                last_activity_at: Some(chrono_now()),
-                ..Default::default()
-              }),
-              persist_op: None,
-            })
-            .await;
-        }
-        "PostToolUseFailure" => {
-          let decision = if is_interrupt.unwrap_or(false) {
-            "denied"
-          } else {
-            "approved"
-          };
-          resolve_pending_approvals_after_tool_outcome(
-            &actor,
-            &persist_tx,
-            &session_id,
-            decision,
-            orbitdock_protocol::WorkStatus::Working,
-          )
-          .await;
-
-          let _ = persist_tx
-            .send(PersistCommand::ClaudeToolIncrement {
-              id: session_id.clone(),
-            })
-            .await;
-          {
-            use crate::runtime::session_state_transitions::{
-              attention_reason_for_status, transition_work_status,
-            };
-            actor
-              .send(SessionCommand::ProcessEvent {
-                event: Input::AttentionUpdated {
-                  attention_reason: attention_reason_for_status(
-                    orbitdock_protocol::WorkStatus::Working,
-                  ),
-                  last_tool: None,
-                  pending_tool_name: None,
-                  pending_tool_input: None,
-                  pending_question: None,
-                },
-              })
-              .await;
-            transition_work_status(
-              &actor,
-              &session_id,
-              orbitdock_protocol::WorkStatus::Working,
-              None,
-            )
-            .await;
-          }
-        }
-        "PermissionRequest" => {
-          let serialized_input = tool_input
-            .as_ref()
-            .and_then(|value| serde_json::to_string(value).ok());
-          let (approval_type, work_status) = classify_permission_request(&tool_name);
-          let request_id =
-            claude_permission_request_id(Some(&actor), &tool_name, tool_use_id.as_deref());
-          let fallback_question = extract_question_from_tool_input(tool_input.as_ref());
-          let question_text =
-            approval_question(serialized_input.as_deref(), fallback_question.as_deref());
-          let plan_text = extract_plan_from_tool_input(tool_input.as_ref());
-          let snapshot = actor.snapshot();
-          let is_duplicate_request = permission_request_matches_snapshot(
-            &snapshot,
-            &PermissionRequestSnapshotMatch {
-              request_id: request_id.as_str(),
-              tool_name: tool_name.as_str(),
-              tool_input: serialized_input.as_deref(),
-              question: question_text.as_deref(),
-              work_status,
-              permission_mode: permission_mode.as_deref(),
-              plan_text: plan_text.as_deref(),
-            },
-          );
-
-          if is_duplicate_request {
-            // Duplicate permission request with unchanged effective state — skip.
-          } else {
-            // Set last_tool before the approval
-            actor
-              .send(SessionCommand::ProcessEvent {
-                event: Input::AttentionUpdated {
-                  attention_reason:
-                    crate::runtime::session_state_transitions::attention_reason_for_status(
-                      work_status,
-                    ),
-                  last_tool: Some(tool_name.clone()),
-                  pending_tool_name: None,
-                  pending_tool_input: None,
-                  pending_question: None,
-                },
-              })
-              .await;
-
-            // Handle plan update if present
-            if let Some(plan_text) = plan_text {
-              actor
-                .send(SessionCommand::ProcessEvent {
-                  event: Input::PlanUpdated(plan_text),
-                })
-                .await;
-            }
-
-            // ApprovalRequested handles: pending fields, work_status persist,
-            // approval row persist, and broadcast — all atomically.
-            actor
-              .send(SessionCommand::ProcessEvent {
-                event: Input::ApprovalRequested {
-                  request_id: request_id.clone(),
-                  approval_type,
-                  tool_name: Some(tool_name.clone()),
-                  tool_input: serialized_input.clone(),
-                  command: None,
-                  file_path: None,
-                  diff: None,
-                  question: question_text.clone(),
-                  permission_reason: None,
-                  requested_permissions: None,
-                  proposed_amendment: None,
-                  permission_suggestions: permission_suggestions.clone(),
-                  elicitation_mode: None,
-                  elicitation_schema: None,
-                  elicitation_url: None,
-                  elicitation_message: None,
-                  mcp_server_name: None,
-                  network_host: None,
-                  network_protocol: None,
-                },
-              })
-              .await;
-
-            if let Some(ref pm) = permission_mode {
-              actor
-                .send(SessionCommand::ProcessEvent {
-                  event: Input::PermissionModeChanged { mode: pm.clone() },
-                })
-                .await;
-            }
-
-            crate::runtime::session_registry::flush_and_publish_conversation(
-              &persist_tx,
-              state,
-              &session_id,
-            )
-            .await;
-          }
-        }
-        _ => {}
-      }
-
-      maybe_sync_transcript_messages(&actor, &persist_tx, &options).await;
+      handle_claude_tool_event(
+        state,
+        &options,
+        session_id,
+        cwd,
+        hook_event_name,
+        tool_name,
+        tool_input,
+        tool_use_id,
+        permission_suggestions,
+        is_interrupt,
+        permission_mode,
+      )
+      .await;
     }
 
     ClientMessage::ClaudeSubagentEvent {
@@ -1130,205 +171,15 @@ pub async fn handle_hook_message_with_options(
       stop_hook_active: _,
       last_assistant_message: _,
     } => {
-      match resolve_claude_hook_routing(state, &session_id).await {
-        ClaudeHookRoutingDecision::ManagedDirect { owner_session_id } => {
-          cleanup_claude_shadow_session(state, &session_id, "managed_direct_session").await;
-          let persist_tx = state.persist().clone();
-
-          match hook_event_name.as_str() {
-            "SubagentStart" => {
-              let raw_type = agent_type.as_deref().unwrap_or("unknown");
-              let normalized_type = AgentType::from_str_normalized(raw_type);
-              let _ = persist_tx
-                .send(PersistCommand::ClaudeSubagentStart {
-                  id: agent_id.clone(),
-                  session_id: owner_session_id.clone(),
-                  agent_type: normalized_type.clone(),
-                })
-                .await;
-              publish_claude_subagent_update(
-                state,
-                &owner_session_id,
-                ClaudeSubagentUpdate::Started {
-                  agent_id: agent_id.clone(),
-                  agent_type: normalized_type.clone(),
-                },
-              )
-              .await;
-              let _ = persist_tx
-                .send(PersistCommand::ClaudeSessionUpdate {
-                  id: owner_session_id,
-                  work_status: None,
-                  attention_reason: None,
-                  last_tool: None,
-                  last_tool_at: None,
-                  pending_tool_name: None,
-                  pending_tool_input: None,
-                  pending_question: None,
-                  source: None,
-                  agent_type: None,
-                  permission_mode: None,
-                  active_subagent_id: Some(Some(agent_id)),
-                  active_subagent_type: Some(Some(normalized_type)),
-                  first_prompt: None,
-                  compact_count_increment: false,
-                })
-                .await;
-            }
-            "SubagentStop" => {
-              let subagent_id = agent_id.clone();
-              let _ = persist_tx
-                .send(PersistCommand::ClaudeSubagentEnd {
-                  id: subagent_id.clone(),
-                  transcript_path: agent_transcript_path,
-                })
-                .await;
-              publish_claude_subagent_update(
-                state,
-                &owner_session_id,
-                ClaudeSubagentUpdate::Stopped {
-                  agent_id: subagent_id,
-                },
-              )
-              .await;
-              let _ = persist_tx
-                .send(PersistCommand::ClaudeSessionUpdate {
-                  id: owner_session_id,
-                  work_status: None,
-                  attention_reason: None,
-                  last_tool: None,
-                  last_tool_at: None,
-                  pending_tool_name: None,
-                  pending_tool_input: None,
-                  pending_question: None,
-                  source: None,
-                  agent_type: None,
-                  permission_mode: None,
-                  active_subagent_id: Some(None),
-                  active_subagent_type: Some(None),
-                  first_prompt: None,
-                  compact_count_increment: false,
-                })
-                .await;
-            }
-            _ => {}
-          }
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreShadowedByDirect => {
-          cleanup_claude_shadow_session(state, &session_id, "direct_owner_exists").await;
-          return;
-        }
-        ClaudeHookRoutingDecision::IgnoreOwnershipLookupFailed => {
-          return;
-        }
-        ClaudeHookRoutingDecision::Passive => {}
-      }
-
-      let persist_tx = state.persist().clone();
-
-      // If session doesn't exist yet, try to materialize from pending cache.
-      // Subagent events don't carry cwd, so peek it from the pending entry.
-      if state.get_session(&session_id).is_none() {
-        if let Some(pending_cwd) = state.peek_pending_hook_cwd(Provider::Claude, &session_id) {
-          let git_info = crate::domain::git::repo::resolve_git_info(&pending_cwd).await;
-          let derived_tp = claude_transcript_path_from_cwd(&pending_cwd, &session_id);
-          materialize_claude_session(
-            &session_id,
-            &pending_cwd,
-            derived_tp,
-            git_info.as_ref(),
-            state,
-            &persist_tx,
-          )
-          .await;
-        }
-      }
-
-      if let Some(existing) = state.get_session(&session_id) {
-        if existing.snapshot().provider == Provider::Codex {
-          return;
-        }
-      }
-
-      match hook_event_name.as_str() {
-        "SubagentStart" => {
-          let raw_type = agent_type.as_deref().unwrap_or("unknown");
-          let normalized_type = AgentType::from_str_normalized(raw_type);
-          let _ = persist_tx
-            .send(PersistCommand::ClaudeSubagentStart {
-              id: agent_id.clone(),
-              session_id: session_id.clone(),
-              agent_type: normalized_type.clone(),
-            })
-            .await;
-          publish_claude_subagent_update(
-            state,
-            &session_id,
-            ClaudeSubagentUpdate::Started {
-              agent_id: agent_id.clone(),
-              agent_type: normalized_type.clone(),
-            },
-          )
-          .await;
-          let _ = persist_tx
-            .send(PersistCommand::ClaudeSessionUpdate {
-              id: session_id,
-              work_status: None,
-              attention_reason: None,
-              last_tool: None,
-              last_tool_at: None,
-              pending_tool_name: None,
-              pending_tool_input: None,
-              pending_question: None,
-              source: None,
-              agent_type: None,
-              permission_mode: None,
-              active_subagent_id: Some(Some(agent_id)),
-              active_subagent_type: Some(Some(normalized_type)),
-              first_prompt: None,
-              compact_count_increment: false,
-            })
-            .await;
-        }
-        "SubagentStop" => {
-          let subagent_id = agent_id.clone();
-          let _ = persist_tx
-            .send(PersistCommand::ClaudeSubagentEnd {
-              id: subagent_id.clone(),
-              transcript_path: agent_transcript_path,
-            })
-            .await;
-          publish_claude_subagent_update(
-            state,
-            &session_id,
-            ClaudeSubagentUpdate::Stopped {
-              agent_id: subagent_id,
-            },
-          )
-          .await;
-          let _ = persist_tx
-            .send(PersistCommand::ClaudeSessionUpdate {
-              id: session_id,
-              work_status: None,
-              attention_reason: None,
-              last_tool: None,
-              last_tool_at: None,
-              pending_tool_name: None,
-              pending_tool_input: None,
-              pending_question: None,
-              source: None,
-              agent_type: None,
-              permission_mode: None,
-              active_subagent_id: Some(None),
-              active_subagent_type: Some(None),
-              first_prompt: None,
-              compact_count_increment: false,
-            })
-            .await;
-        }
-        _ => {}
-      }
+      handle_claude_subagent_event(
+        state,
+        session_id,
+        hook_event_name,
+        agent_id,
+        agent_type,
+        agent_transcript_path,
+      )
+      .await;
     }
 
     _ => {
