+++ orbitdock-server/crates/server/src/domain/mission_control/config.rs
@@ -1,141 +1,15 @@
-use anyhow::{Context, Result};
-use serde::Deserialize;
-
 #[path = "config_model.rs"]
 mod config_model;
+mod migration;
+mod parser;
+mod scaffold;
+mod serializer;
 
 pub use self::config_model::*;
-use super::template::default_mission_template;
-use super::tracker::TrackerConfig;
-
-// ── Symphony WORKFLOW.md migration ───────────────────────────────────
-
-/// Symphony's WORKFLOW.md tracker config.
-#[derive(Debug, Clone, Deserialize, Default)]
-struct SymphonyTracker {
-  #[serde(default)]
-  kind: String,
-  #[serde(default)]
-  project_slug: Option<String>,
-  #[serde(default)]
-  active_states: Vec<String>,
-}
-
-/// Symphony's WORKFLOW.md polling config.
-#[derive(Debug, Clone, Deserialize, Default)]
-struct SymphonyPolling {
-  #[serde(default)]
-  interval_ms: u64,
-}
-
-/// Symphony's WORKFLOW.md agent config.
-#[derive(Debug, Clone, Deserialize, Default)]
-struct SymphonyAgent {
-  #[serde(default)]
-  max_concurrent_agents: u32,
-}
-
-/// Symphony's WORKFLOW.md codex config.
-#[derive(Debug, Clone, Deserialize, Default)]
-struct SymphonyCodex {
-  #[serde(default)]
-  command: Option<String>,
-}
-
-/// Top-level Symphony WORKFLOW.md schema.
-#[derive(Debug, Clone, Deserialize, Default)]
-struct SymphonyWorkflow {
-  #[serde(default)]
-  tracker: SymphonyTracker,
-  #[serde(default)]
-  polling: SymphonyPolling,
-  #[serde(default)]
-  agent: SymphonyAgent,
-  #[serde(default)]
-  codex: SymphonyCodex,
-}
-
-/// Try to parse a Symphony WORKFLOW.md and convert to MissionConfig.
-/// Returns None if the content doesn't look like a Symphony workflow.
-pub fn try_parse_symphony_workflow(content: &str) -> Option<MissionConfig> {
-  let content = content.trim();
-  if !content.starts_with("---") {
-    return None;
-  }
-
-  let after_first = &content[3..];
-  let end_idx = after_first.find("\n---")?;
-  let yaml_block = &after_first[..end_idx];
-
-  // Must have at least one Symphony-specific key
-  if !yaml_block.contains("tracker:")
-    && !yaml_block.contains("polling:")
-    && !yaml_block.contains("agent:")
-    && !yaml_block.contains("codex:")
-  {
-    return None;
-  }
-
-  let symphony: SymphonyWorkflow = serde_yaml::from_str(yaml_block).ok()?;
-
-  // Only convert if there's meaningful config (not all defaults)
-  let has_config = !symphony.tracker.kind.is_empty()
-    || symphony.tracker.project_slug.is_some()
-    || !symphony.tracker.active_states.is_empty()
-    || symphony.polling.interval_ms > 0
-    || symphony.agent.max_concurrent_agents > 0;
-
-  if !has_config {
-    return None;
-  }
-
-  // Detect provider from codex command
-  let primary = if symphony.codex.command.is_some() {
-    "codex".to_string()
-  } else {
-    "claude".to_string()
-  };
-
-  let interval = if symphony.polling.interval_ms > 0 {
-    symphony.polling.interval_ms / 1000
-  } else {
-    60
-  };
-
-  let max_concurrent = if symphony.agent.max_concurrent_agents > 0 {
-    symphony.agent.max_concurrent_agents
-  } else {
-    3
-  };
-
-  Some(MissionConfig {
-    tracker: if symphony.tracker.kind.is_empty() {
-      "linear".to_string()
-    } else {
-      symphony.tracker.kind
-    },
-    provider: ProviderConfig {
-      strategy: "single".to_string(),
-      primary,
-      secondary: None,
-      max_concurrent,
-      max_concurrent_primary: None,
-    },
-    agent: AgentConfig::default(),
-    workspace: WorkspaceConfig::default(),
-    trigger: TriggerConfig {
-      kind: "polling".to_string(),
-      interval,
-      filters: TriggerFilters {
-        labels: Vec::new(),
-        states: symphony.tracker.active_states,
-        project: symphony.tracker.project_slug,
-        team: None,
-      },
-    },
-    orchestration: OrchestrationConfig::default(),
-  })
-}
+pub use self::migration::{migrate_workflow_content, try_parse_symphony_workflow};
+pub use self::parser::parse_mission_file;
+pub use self::scaffold::generate_scaffold;
+pub use self::serializer::{serialize_mission_file, serialize_mission_file_preserving};
 
 // ── Public API ───────────────────────────────────────────────────────
 
@@ -144,396 +18,6 @@ pub fn try_parse_symphony_workflow(content: &str) -> Option<MissionConfig> {
 pub struct MissionDefinition {
   pub config: MissionConfig,
   pub prompt_template: String,
-}
-
-/// Parse a MISSION.md file: YAML front matter between `---` fences, rest is Liquid template.
-///
-/// Supports the top-level `MissionConfig` schema and legacy flat schema for backward compat.
-pub fn parse_mission_file(content: &str) -> Result<MissionDefinition> {
-  let content = content.trim();
-  if !content.starts_with("---") {
-    anyhow::bail!("MISSION.md must start with YAML front matter (---)");
-  }
-
-  let after_first = &content[3..];
-  let end_idx = after_first
-    .find("\n---")
-    .context("MISSION.md missing closing --- for YAML front matter")?;
-
-  let yaml_block = &after_first[..end_idx];
-  let prompt_start = 3 + end_idx + 4; // skip past "\n---"
-  let prompt_template = if prompt_start < content.len() {
-    content[prompt_start..].trim().to_string()
-  } else {
-    String::new()
-  };
-
-  // Try parsing as MissionConfig directly (top-level keys).
-  // This can fail if the YAML has legacy fields (e.g. `provider: "claude"` string
-  // instead of `provider: {strategy: ...}` struct), so we fall back to legacy.
-  if let Ok(config) = serde_yaml::from_str::<MissionConfig>(yaml_block) {
-    // Check if parsed config differs from defaults — if it's all defaults,
-    // the YAML may not contain any recognized Mission Control fields.
-    let defaults = MissionConfig::default();
-    let has_mission_config = config.tracker != defaults.tracker
-      || config.provider.strategy != defaults.provider.strategy
-      || config.provider.primary != defaults.provider.primary
-      || config.provider.secondary.is_some()
-      || config.provider.max_concurrent != defaults.provider.max_concurrent
-      || config.provider.max_concurrent_primary.is_some()
-      || !config.workspace.is_empty()
-      || config.trigger.kind != defaults.trigger.kind
-      || config.trigger.interval != defaults.trigger.interval
-      || !config.trigger.filters.labels.is_empty()
-      || !config.trigger.filters.states.is_empty()
-      || config.trigger.filters.project.is_some()
-      || config.trigger.filters.team.is_some()
-      || config.orchestration.max_retries != defaults.orchestration.max_retries
-      || config.orchestration.stall_timeout != defaults.orchestration.stall_timeout
-      || config.orchestration.base_branch != defaults.orchestration.base_branch;
-
-    if has_mission_config {
-      return Ok(MissionDefinition {
-        config,
-        prompt_template,
-      });
-    }
-
-    // MissionConfig parsed but is all defaults — check if the YAML has any
-    // recognized MissionConfig top-level keys (even with default values).
-    let has_recognized_keys = yaml_block.contains("tracker:")
-      || yaml_block.contains("provider:")
-      || yaml_block.contains("workspace:")
-      || yaml_block.contains("agent:")
-      || yaml_block.contains("trigger:")
-      || yaml_block.contains("orchestration:");
-
-    if has_recognized_keys {
-      return Ok(MissionDefinition {
-        config,
-        prompt_template,
-      });
-    }
-  }
-
-  anyhow::bail!(
-    "MISSION.md does not contain OrbitDock mission configuration. \
-         Use 'Generate MISSION.md' to create one."
-  )
-}
-
-/// Reconstruct MISSION.md content from config + prompt template.
-///
-/// Writes config keys at the top level of the YAML front matter.
-pub fn serialize_mission_file(config: &MissionConfig, prompt_template: &str) -> Result<String> {
-  let yaml = serde_yaml::to_string(config).context("serialize config to YAML")?;
-  Ok(format!("---\n{}---\n\n{}", yaml, prompt_template))
-}
-
-/// Serialize config while preserving non-mission content from an existing MISSION.md.
-///
-/// - If `existing_content` has YAML front matter with non-mission keys, they are preserved.
-/// - If `prompt_template` is empty, the existing body (text after front matter) is kept.
-/// - If `existing_content` has no front matter, the mission config is prepended and
-///   the existing content becomes the prompt body.
-pub fn serialize_mission_file_preserving(
-  config: &MissionConfig,
-  prompt_template: &str,
-  existing_content: Option<&str>,
-) -> Result<String> {
-  let Some(existing) = existing_content else {
-    return serialize_mission_file(config, prompt_template);
-  };
-
-  let trimmed = existing.trim();
-  if !trimmed.starts_with("---") {
-    // No front matter — prepend mission config
-    let body = if prompt_template.is_empty() {
-      trimmed
-    } else {
-      prompt_template
-    };
-    return serialize_mission_file(config, body);
-  }
-
-  // Has front matter — parse existing YAML, inject/replace mission config keys
-  let after_first = &trimmed[3..];
-  let end_idx = after_first
-    .find("\n---")
-    .context("existing MISSION.md missing closing ---")?;
-  let existing_yaml = &after_first[..end_idx];
-  let existing_body_start = 3 + end_idx + 4;
-  let existing_body = if existing_body_start < trimmed.len() {
-    trimmed[existing_body_start..].trim()
-  } else {
-    ""
-  };
-
-  // Parse existing YAML as generic mapping, inject mission config keys
-  let mut mapping: serde_yaml::Mapping = serde_yaml::from_str(existing_yaml).unwrap_or_default();
-  let config_value =
-    serde_yaml::to_value(config).context("serialize mission config to YAML value")?;
-
-  // Inject each config key directly into the mapping (flat, no wrapper)
-  if let serde_yaml::Value::Mapping(config_map) = config_value {
-    for (k, v) in config_map {
-      mapping.insert(k, v);
-    }
-  }
-
-  let yaml = serde_yaml::to_string(&mapping).context("serialize merged YAML")?;
-
-  // Body: use provided prompt_template if non-empty, else preserve existing body
-  let body = if prompt_template.is_empty() {
-    existing_body
-  } else {
-    prompt_template
-  };
-
-  Ok(format!("---\n{}---\n\n{}", yaml, body))
-}
-
-/// Flat partial-update for mission configuration — mirrors the REST request
-/// but lives in the domain so it can be tested independently.
-#[derive(Default)]
-pub struct MissionConfigUpdate {
-  // Provider
-  pub provider_strategy: Option<String>,
-  pub primary_provider: Option<String>,
-  pub secondary_provider: Option<Option<String>>,
-  pub max_concurrent: Option<u32>,
-  pub max_concurrent_primary: Option<Option<u32>>,
-  // Agent — Claude
-  pub agent_claude_model: Option<Option<String>>,
-  pub agent_claude_effort: Option<Option<String>>,
-  pub agent_claude_permission_mode: Option<Option<String>>,
-  pub agent_claude_allowed_tools: Option<Vec<String>>,
-  pub agent_claude_disallowed_tools: Option<Vec<String>>,
-  pub agent_claude_allow_bypass_permissions: Option<bool>,
-  // Agent — Codex
-  pub agent_codex_model: Option<Option<String>>,
-  pub agent_codex_effort: Option<Option<String>>,
-  pub agent_codex_approval_policy: Option<Option<String>>,
-  pub agent_codex_sandbox_mode: Option<Option<String>>,
-  pub agent_codex_collaboration_mode: Option<Option<String>>,
-  pub agent_codex_multi_agent: Option<Option<bool>>,
-  pub agent_codex_personality: Option<Option<String>>,
-  pub agent_codex_service_tier: Option<Option<String>>,
-  pub agent_codex_developer_instructions: Option<Option<String>>,
-  // Trigger
-  pub trigger_kind: Option<String>,
-  pub poll_interval: Option<u64>,
-  pub label_filter: Option<Vec<String>>,
-  pub state_filter: Option<Vec<String>>,
-  pub project_key: Option<Option<String>>,
-  pub team_key: Option<Option<String>>,
-  // Orchestration
-  pub max_retries: Option<u32>,
-  pub stall_timeout: Option<u64>,
-  pub base_branch: Option<String>,
-  pub worktree_root_dir: Option<Option<String>>,
-  pub state_on_dispatch: Option<String>,
-  pub state_on_complete: Option<String>,
-  // Tracker
-  pub tracker: Option<String>,
-}
-
-impl MissionConfig {
-  /// Apply a partial update, merging only the fields that are `Some`.
-  pub fn apply_update(&mut self, u: MissionConfigUpdate) {
-    // Provider
-    if let Some(v) = u.provider_strategy {
-      self.provider.strategy = v;
-    }
-    if let Some(v) = u.primary_provider {
-      self.provider.primary = v;
-    }
-    if let Some(v) = u.secondary_provider {
-      self.provider.secondary = v;
-    }
-    if let Some(v) = u.max_concurrent {
-      self.provider.max_concurrent = v;
-    }
-    if let Some(v) = u.max_concurrent_primary {
-      self.provider.max_concurrent_primary = v;
-    }
-
-    // Agent — Claude
-    let has_claude = u.agent_claude_model.is_some()
-      || u.agent_claude_effort.is_some()
-      || u.agent_claude_permission_mode.is_some()
-      || u.agent_claude_allowed_tools.is_some()
-      || u.agent_claude_disallowed_tools.is_some()
-      || u.agent_claude_allow_bypass_permissions.is_some();
-    if has_claude {
-      let claude = self
-        .agent
-        .claude
-        .get_or_insert_with(ClaudeAgentConfig::default);
-      if let Some(v) = u.agent_claude_model {
-        claude.model = v;
-      }
-      if let Some(v) = u.agent_claude_effort {
-        claude.effort = v;
-      }
-      if let Some(v) = u.agent_claude_permission_mode {
-        claude.permission_mode = v;
-      }
-      if let Some(v) = u.agent_claude_allowed_tools {
-        claude.allowed_tools = v;
-      }
-      if let Some(v) = u.agent_claude_disallowed_tools {
-        claude.disallowed_tools = v;
-      }
-      if let Some(v) = u.agent_claude_allow_bypass_permissions {
-        claude.allow_bypass_permissions = v;
-      }
-    }
-
-    // Agent — Codex
-    let has_codex = u.agent_codex_model.is_some()
-      || u.agent_codex_effort.is_some()
-      || u.agent_codex_approval_policy.is_some()
-      || u.agent_codex_sandbox_mode.is_some()
-      || u.agent_codex_collaboration_mode.is_some()
-      || u.agent_codex_multi_agent.is_some()
-      || u.agent_codex_personality.is_some()
-      || u.agent_codex_service_tier.is_some()
-      || u.agent_codex_developer_instructions.is_some();
-    if has_codex {
-      let codex = self
-        .agent
-        .codex
-        .get_or_insert_with(CodexAgentConfig::default);
-      if let Some(v) = u.agent_codex_model {
-        codex.model = v;
-      }
-      if let Some(v) = u.agent_codex_effort {
-        codex.effort = v;
-      }
-      if let Some(v) = u.agent_codex_approval_policy {
-        codex.approval_policy = v;
-      }
-      if let Some(v) = u.agent_codex_sandbox_mode {
-        codex.sandbox_mode = v;
-      }
-      if let Some(v) = u.agent_codex_collaboration_mode {
-        codex.collaboration_mode = v;
-      }
-      if let Some(v) = u.agent_codex_multi_agent {
-        codex.multi_agent = v;
-      }
-      if let Some(v) = u.agent_codex_personality {
-        codex.personality = v;
-      }
-      if let Some(v) = u.agent_codex_service_tier {
-        codex.service_tier = v;
-      }
-      if let Some(v) = u.agent_codex_developer_instructions {
-        codex.developer_instructions = v;
-      }
-    }
-
-    // Trigger
-    if let Some(v) = u.trigger_kind {
-      self.trigger.kind = v;
-    }
-    if let Some(v) = u.poll_interval {
-      self.trigger.interval = v;
-    }
-    if let Some(v) = u.label_filter {
-      self.trigger.filters.labels = v;
-    }
-    if let Some(v) = u.state_filter {
-      self.trigger.filters.states = v;
-    }
-    if let Some(v) = u.project_key {
-      self.trigger.filters.project = v;
-    }
-    if let Some(v) = u.team_key {
-      self.trigger.filters.team = v;
-    }
-
-    // Orchestration
-    if let Some(v) = u.max_retries {
-      self.orchestration.max_retries = v;
-    }
-    if let Some(v) = u.stall_timeout {
-      self.orchestration.stall_timeout = v;
-    }
-    if let Some(v) = u.base_branch {
-      self.orchestration.base_branch = v;
-    }
-    if let Some(v) = u.worktree_root_dir {
-      self.orchestration.worktree_root_dir = v;
-    }
-    if let Some(v) = u.state_on_dispatch {
-      self.orchestration.state_on_dispatch = v;
-    }
-    if let Some(v) = u.state_on_complete {
-      self.orchestration.state_on_complete = v;
-    }
-
-    // Tracker
-    if let Some(v) = u.tracker {
-      self.tracker = v;
-    }
-  }
-}
-
-// ── Scaffold & migration ─────────────────────────────────────────────
-
-/// Generate a scaffold MISSION.md for a given provider.
-///
-/// Returns `(file_content, parsed_config, prompt_template)`.
-pub fn generate_scaffold(provider: &str, tracker: &str) -> Result<(String, MissionConfig, String)> {
-  let file_content = default_mission_template(provider, tracker);
-  let parsed = parse_mission_file(&file_content).context("parse scaffolded template")?;
-  let prompt_template = parsed.prompt_template.clone();
-  Ok((file_content, parsed.config, prompt_template))
-}
-
-/// Convert a Symphony WORKFLOW.md content string into MISSION.md format.
-///
-/// Returns `(file_content, parsed_config, prompt_template)`.
-pub fn migrate_workflow_content(
-  workflow_content: &str,
-  fallback_provider: &str,
-) -> Result<(String, MissionConfig, String)> {
-  let config = try_parse_symphony_workflow(workflow_content)
-    .context("WORKFLOW.md does not contain recognized Symphony configuration")?;
-
-  // Extract body from the WORKFLOW.md (everything after the YAML front matter).
-  let prompt_template = {
-    let trimmed = workflow_content.trim();
-    let body = if let Some(after_first) = trimmed.strip_prefix("---") {
-      if let Some(end_idx) = after_first.find("\n---") {
-        let prompt_start = end_idx + 4; // skip past "\n---"
-        if prompt_start < after_first.len() {
-          after_first[prompt_start..].trim()
-        } else {
-          ""
-        }
-      } else {
-        ""
-      }
-    } else {
-      ""
-    };
-
-    if body.is_empty() {
-      let full_template = default_mission_template(fallback_provider, &config.tracker);
-      parse_mission_file(&full_template)
-        .map(|def| def.prompt_template)
-        .unwrap_or_default()
-    } else {
-      body.to_string()
-    }
-  };
-
-  let file_content =
-    serialize_mission_file(&config, &prompt_template).context("serialize migrated MISSION.md")?;
-  Ok((file_content, config, prompt_template))
 }
 
 #[cfg(test)]
