+++ orbitdock-server/crates/server/src/domain/mission_control/config_model.rs
@@ -68,6 +68,50 @@ impl Default for MissionConfig {
   }
 }
 
+/// Flat partial-update for mission configuration.
+#[derive(Default)]
+pub struct MissionConfigUpdate {
+  // Provider
+  pub provider_strategy: Option<String>,
+  pub primary_provider: Option<String>,
+  pub secondary_provider: Option<Option<String>>,
+  pub max_concurrent: Option<u32>,
+  pub max_concurrent_primary: Option<Option<u32>>,
+  // Agent — Claude
+  pub agent_claude_model: Option<Option<String>>,
+  pub agent_claude_effort: Option<Option<String>>,
+  pub agent_claude_permission_mode: Option<Option<String>>,
+  pub agent_claude_allowed_tools: Option<Vec<String>>,
+  pub agent_claude_disallowed_tools: Option<Vec<String>>,
+  pub agent_claude_allow_bypass_permissions: Option<bool>,
+  // Agent — Codex
+  pub agent_codex_model: Option<Option<String>>,
+  pub agent_codex_effort: Option<Option<String>>,
+  pub agent_codex_approval_policy: Option<Option<String>>,
+  pub agent_codex_sandbox_mode: Option<Option<String>>,
+  pub agent_codex_collaboration_mode: Option<Option<String>>,
+  pub agent_codex_multi_agent: Option<Option<bool>>,
+  pub agent_codex_personality: Option<Option<String>>,
+  pub agent_codex_service_tier: Option<Option<String>>,
+  pub agent_codex_developer_instructions: Option<Option<String>>,
+  // Trigger
+  pub trigger_kind: Option<String>,
+  pub poll_interval: Option<u64>,
+  pub label_filter: Option<Vec<String>>,
+  pub state_filter: Option<Vec<String>>,
+  pub project_key: Option<Option<String>>,
+  pub team_key: Option<Option<String>>,
+  // Orchestration
+  pub max_retries: Option<u32>,
+  pub stall_timeout: Option<u64>,
+  pub base_branch: Option<String>,
+  pub worktree_root_dir: Option<Option<String>>,
+  pub state_on_dispatch: Option<String>,
+  pub state_on_complete: Option<String>,
+  // Tracker
+  pub tracker: Option<String>,
+}
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(default)]
 pub struct ProviderConfig {
@@ -308,50 +352,6 @@ pub struct ResolvedAgentSettings {
 }
 
 impl AgentConfig {
-  /// Resolve agent settings for the given provider name.
-  ///
-  /// Mission agents run headless - defaults ensure agents can operate
-  /// autonomously without stalling on permission prompts.
-  pub fn resolve_settings(
-    &self,
-    provider: &str,
-    defaults: Option<ResolvedAgentSettings>,
-  ) -> ResolvedAgentSettings {
-    let mut settings = defaults.unwrap_or_default();
-    match provider {
-      "claude" => {
-        if let Some(config) = self.claude.as_ref() {
-          settings.model = config.model.clone().or(settings.model);
-          settings.effort = config.effort.clone().or(settings.effort);
-          settings.permission_mode = config.permission_mode.clone().or(settings.permission_mode);
-          settings.allow_bypass_permissions = config.allow_bypass_permissions;
-        }
-      }
-      "codex" => {
-        if let Some(config) = self.codex.as_ref() {
-          settings.model = config.model.clone().or(settings.model);
-          settings.effort = config.effort.clone().or(settings.effort);
-          settings.approval_policy = config.approval_policy.clone().or(settings.approval_policy);
-          settings.sandbox_mode = config.sandbox_mode.clone().or(settings.sandbox_mode);
-          settings.collaboration_mode = config
-            .collaboration_mode
-            .clone()
-            .or(settings.collaboration_mode);
-          settings.multi_agent = config.multi_agent.or(settings.multi_agent);
-          settings.personality = config.personality.clone().or(settings.personality);
-          settings.service_tier = config.service_tier.clone().or(settings.service_tier);
-          settings.developer_instructions = config
-            .developer_instructions
-            .clone()
-            .or(settings.developer_instructions);
-          settings.model_provider = config.model_provider.clone().or(settings.model_provider);
-        }
-      }
-      _ => {}
-    }
-    settings
-  }
-
   /// Resolve mission runtime settings for the given provider name.
   pub fn resolve_for_provider(&self, provider: &str) -> ResolvedAgentSettings {
     match provider {
@@ -432,6 +432,147 @@ impl MissionConfig {
       team_key: self.trigger.filters.team.clone(),
       label_filter: self.trigger.filters.labels.clone(),
       state_filter: self.trigger.filters.states.clone(),
+    }
+  }
+
+  /// Apply a partial update, merging only the fields that are `Some`.
+  pub fn apply_update(&mut self, u: MissionConfigUpdate) {
+    // Provider
+    if let Some(v) = u.provider_strategy {
+      self.provider.strategy = v;
+    }
+    if let Some(v) = u.primary_provider {
+      self.provider.primary = v;
+    }
+    if let Some(v) = u.secondary_provider {
+      self.provider.secondary = v;
+    }
+    if let Some(v) = u.max_concurrent {
+      self.provider.max_concurrent = v;
+    }
+    if let Some(v) = u.max_concurrent_primary {
+      self.provider.max_concurrent_primary = v;
+    }
+
+    // Agent — Claude
+    let has_claude = u.agent_claude_model.is_some()
+      || u.agent_claude_effort.is_some()
+      || u.agent_claude_permission_mode.is_some()
+      || u.agent_claude_allowed_tools.is_some()
+      || u.agent_claude_disallowed_tools.is_some()
+      || u.agent_claude_allow_bypass_permissions.is_some();
+    if has_claude {
+      let claude = self
+        .agent
+        .claude
+        .get_or_insert_with(ClaudeAgentConfig::default);
+      if let Some(v) = u.agent_claude_model {
+        claude.model = v;
+      }
+      if let Some(v) = u.agent_claude_effort {
+        claude.effort = v;
+      }
+      if let Some(v) = u.agent_claude_permission_mode {
+        claude.permission_mode = v;
+      }
+      if let Some(v) = u.agent_claude_allowed_tools {
+        claude.allowed_tools = v;
+      }
+      if let Some(v) = u.agent_claude_disallowed_tools {
+        claude.disallowed_tools = v;
+      }
+      if let Some(v) = u.agent_claude_allow_bypass_permissions {
+        claude.allow_bypass_permissions = v;
+      }
+    }
+
+    // Agent — Codex
+    let has_codex = u.agent_codex_model.is_some()
+      || u.agent_codex_effort.is_some()
+      || u.agent_codex_approval_policy.is_some()
+      || u.agent_codex_sandbox_mode.is_some()
+      || u.agent_codex_collaboration_mode.is_some()
+      || u.agent_codex_multi_agent.is_some()
+      || u.agent_codex_personality.is_some()
+      || u.agent_codex_service_tier.is_some()
+      || u.agent_codex_developer_instructions.is_some();
+    if has_codex {
+      let codex = self
+        .agent
+        .codex
+        .get_or_insert_with(CodexAgentConfig::default);
+      if let Some(v) = u.agent_codex_model {
+        codex.model = v;
+      }
+      if let Some(v) = u.agent_codex_effort {
+        codex.effort = v;
+      }
+      if let Some(v) = u.agent_codex_approval_policy {
+        codex.approval_policy = v;
+      }
+      if let Some(v) = u.agent_codex_sandbox_mode {
+        codex.sandbox_mode = v;
+      }
+      if let Some(v) = u.agent_codex_collaboration_mode {
+        codex.collaboration_mode = v;
+      }
+      if let Some(v) = u.agent_codex_multi_agent {
+        codex.multi_agent = v;
+      }
+      if let Some(v) = u.agent_codex_personality {
+        codex.personality = v;
+      }
+      if let Some(v) = u.agent_codex_service_tier {
+        codex.service_tier = v;
+      }
+      if let Some(v) = u.agent_codex_developer_instructions {
+        codex.developer_instructions = v;
+      }
+    }
+
+    // Trigger
+    if let Some(v) = u.trigger_kind {
+      self.trigger.kind = v;
+    }
+    if let Some(v) = u.poll_interval {
+      self.trigger.interval = v;
+    }
+    if let Some(v) = u.label_filter {
+      self.trigger.filters.labels = v;
+    }
+    if let Some(v) = u.state_filter {
+      self.trigger.filters.states = v;
+    }
+    if let Some(v) = u.project_key {
+      self.trigger.filters.project = v;
+    }
+    if let Some(v) = u.team_key {
+      self.trigger.filters.team = v;
+    }
+
+    // Orchestration
+    if let Some(v) = u.max_retries {
+      self.orchestration.max_retries = v;
+    }
+    if let Some(v) = u.stall_timeout {
+      self.orchestration.stall_timeout = v;
+    }
+    if let Some(v) = u.base_branch {
+      self.orchestration.base_branch = v;
+    }
+    if let Some(v) = u.worktree_root_dir {
+      self.orchestration.worktree_root_dir = v;
+    }
+    if let Some(v) = u.state_on_dispatch {
+      self.orchestration.state_on_dispatch = v;
+    }
+    if let Some(v) = u.state_on_complete {
+      self.orchestration.state_on_complete = v;
+    }
+
+    // Tracker
+    if let Some(v) = u.tracker {
+      self.tracker = v;
     }
   }
 }
